% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blop.R
\name{blop}
\alias{blop}
\alias{bla}
\title{Blop}
\usage{
blop()

bla()
}
\value{
a string
}
\description{
Blop
}
\examples{
blop()
}
\seealso{
Other things: 
\code{\link{ok}()}

Other things: 
\code{\link{ok}()}
}
\concept{things}
